+++ example/posix_uart/pc_clock.h
@@ -18,6 +18,6 @@ typedef struct{
     gmsi_base_t *ptBase;
 }pc_clock_t;
 //int pcclock_Init(pc_clock_t *ptThis, pc_clock_cfg_t *ptCfg);
-int pcclock_Init(uint32_t wObjectAddr, uint32_t wObjectCfgAddr);
-int pcclock_Run(uint32_t wObjectAddr);
+int pcclock_Init(uintptr_t wObjectAddr, uintptr_t wObjectCfgAddr);
+int pcclock_Run(uintptr_t wObjectAddr);
 #endif
